pyproject.toml
[project]
name = "goal-based-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "httpx>=0.28.1",
    "openai-agents>=0.7.0",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
]
